+++ src/resources/customers/index.ts
@@ -11,8 +11,8 @@ import {
   type ListCustomerQueryParameters,
   type UpdateCustomerRequestBody,
 } from './operations/index.js';
-import { CreditBalance, Customer, CustomerCollection } from '../../entities/index.js';
-import { type ICreditBalanceResponse, type ICustomerResponse } from '../../types/index.js';
+import { CreditBalance, Customer, CustomerCollection, AuthToken } from '../../entities/index.js';
+import { type ICreditBalanceResponse, type ICustomerResponse, type IAuthTokenResponse } from '../../types/index.js';
 import { type Response, type ErrorResponse } from '../../internal/index.js';
 
 const CustomerPaths = {
@@ -21,6 +21,7 @@ const CustomerPaths = {
   get: '/customers/{customer_id}',
   update: '/customers/{customer_id}',
   getCustomerBalance: '/customers/{customer_id}/credit-balances',
+  generate: '/customers/{customer_id}/auth-token',
 } as const;
 
 export * from './operations/index.js';
@@ -90,5 +91,20 @@ export class CustomersResource extends BaseResource {
 
   public async archive(customerId: string) {
     return await this.update(customerId, { status: 'archived' });
+  }
+
+  public async generateAuthToken(customerId: string): Promise<AuthToken> {
+    const urlWithPathParams = new PathParameters(CustomerPaths.generate, {
+      customer_id: customerId,
+    }).deriveUrl();
+
+    const response = await this.client.post<undefined, Response<IAuthTokenResponse> | ErrorResponse>(
+      urlWithPathParams,
+      undefined,
+    );
+
+    const data = this.handleResponse<IAuthTokenResponse>(response);
+
+    return new AuthToken(data);
   }
 }