+++ src/types/events/events-response.ts
@@ -11,6 +11,8 @@ import {
   type IBusinessNotificationResponse,
   type ICustomerNotificationResponse,
   type IDiscountNotificationResponse,
+  type IPaymentMethodNotificationResponse,
+  type IPaymentMethodDeletedNotificationResponse,
   type IPayoutNotificationResponse,
   type IPriceNotificationResponse,
   type IProductNotificationResponse,
@@ -82,6 +84,14 @@ interface IDiscountUpdated extends IEventsResponse<IDiscountNotificationResponse
 
 interface IDiscountImported extends IEventsResponse<IDiscountNotificationResponse> {
   event_type: EventName.DiscountImported;
+}
+
+interface IPaymentMethodDeleted extends IEventsResponse<IPaymentMethodDeletedNotificationResponse> {
+  event_type: EventName.PaymentMethodDeleted;
+}
+
+interface IPaymentMethodSaved extends IEventsResponse<IPaymentMethodNotificationResponse> {
+  event_type: EventName.PaymentMethodSaved;
 }
 
 interface IPayoutCreated extends IEventsResponse<IPayoutNotificationResponse> {
@@ -211,6 +221,8 @@ export type IEvents =
   | IDiscountCreated
   | IDiscountUpdated
   | IDiscountImported
+  | IPaymentMethodDeleted
+  | IPaymentMethodSaved
   | IPayoutCreated
   | IPayoutPaid
   | IPriceCreated